refactor: move receive input logic to custom hook
